Remove react, react-dom from peerDependencies
This package does not actually interact directly with either `react` nor `react-dom`. It never imports anything from them or does anything to them. Simply declaring the peerDependency on `react-admin` should be enough.

Removing these peerDependencies helps to install this package into a monorepo where some apps are still on React 17 and others are on 18. Otherwise we have to install using `npm install --force`.